+++ src/pygnssutils/gnssntripclient_cli.py
@@ -26,12 +26,14 @@ from pygnssutils.globals import (
     ENV_NTRIP_PASSWORD,
     ENV_NTRIP_USER,
     EPILOG,
+    MAXCONNECTION,
     NTRIP1,
     NTRIP2,
     OUTPUT_FILE,
     OUTPUT_NONE,
     OUTPUT_SERIAL,
     OUTPUT_SOCKET,
+    OUTPUT_SOCKET_TLS,
 )
 from pygnssutils.gnssntripclient import (
     GGAFIXED,
@@ -96,6 +98,14 @@ def main():
         choices=[0, 1],
         default=0,
     )
+    ap.add_argument(
+        "--selfsign",
+        required=False,
+        help=("Allow self-signed TLS certificate?"),
+        type=int,
+        choices=[0, 1],
+        default=0,
+    ),
     ap.add_argument(
         "-I",
         "--ipprot",
@@ -204,10 +214,17 @@ def main():
             f"CLI output type {OUTPUT_NONE} = none, "
             f"{OUTPUT_FILE} = binary file, "
             f"{OUTPUT_SERIAL} = serial port, "
-            f"{OUTPUT_SOCKET} = TCP socket server"
+            f"{OUTPUT_SOCKET} = TCP socket server, "
+            f"{OUTPUT_SOCKET_TLS} = TCP socket server with TLS"
         ),
         type=int,
-        choices=[OUTPUT_NONE, OUTPUT_FILE, OUTPUT_SERIAL, OUTPUT_SOCKET],
+        choices=[
+            OUTPUT_NONE,
+            OUTPUT_FILE,
+            OUTPUT_SERIAL,
+            OUTPUT_SOCKET,
+            OUTPUT_SOCKET_TLS,
+        ],
         default=OUTPUT_NONE,
     )
     ap.add_argument(
@@ -217,7 +234,8 @@ def main():
             f"Output medium as formatted string. "
             f"If clioutput = {OUTPUT_FILE}, format = file name (e.g. '/home/myuser/rtcm.log'); "
             f"If clioutput = {OUTPUT_SERIAL}, format = port@baudrate (e.g. '/dev/tty.ACM0@38400'); "
-            f"If clioutput = {OUTPUT_SOCKET}, format = hostip:port (e.g. '0.0.0.0:50010'). "
+            f"If clioutput = {OUTPUT_SOCKET} or {OUTPUT_SOCKET_TLS}, "
+            "format = hostip:port (e.g. '0.0.0.0:50010'). "
             "NB: gnssntripclient will have exclusive use of any serial or server port."
         ),
         default=None,
@@ -237,14 +255,15 @@ def main():
             with Serial(port, int(baud), timeout=3) as output:
                 kwargs["output"] = output
                 runclient(**kwargs)
-        elif cliout == OUTPUT_SOCKET:
+        elif cliout in (OUTPUT_SOCKET, OUTPUT_SOCKET_TLS):
             host, port = kwargs["output"].split(":")
+            tls = cliout == OUTPUT_SOCKET_TLS
             kwargs["output"] = Queue()
             # socket server runs as background thread, piping
             # output from ntrip client via a message queue
             Thread(
                 target=runserver,
-                args=(host, int(port), kwargs["output"]),
+                args=(host, int(port), kwargs["output"], 0, MAXCONNECTION, tls, NTRIP2),
                 daemon=True,
             ).start()
             runclient(**kwargs)